move load config
